move bitoperation-related components to new folder
