+++ pages/logical-operations.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import {Component}        from 'react';
-import LogicalOperations  from '../components/LogicalOperations';
+import LogicalOperations  from '../components/logical-operations/LogicalOperations';
 
 
 export default class LogicalOperationsPage extends Component {
